SSCSSI-247: SubmitAppealService - case create or update from SYA
