+++ src/main/java/uk/gov/hmcts/reform/sscs/service/SubmitAppealServiceBase.java
@@ -10,13 +10,12 @@ import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.NON_COMPLIANT;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.EventType.VALID_APPEAL_CREATED;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.State.READY_TO_LIST;
 import static uk.gov.hmcts.reform.sscs.service.RegionalProcessingCenterService.getFirstHalfOfPostcode;
-import static uk.gov.hmcts.reform.sscs.transform.deserialize.SubmitYourAppealToCcdCaseDataDeserializer.convertSyaToCcdCaseDataV1;
 
 import feign.FeignException;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.function.Function;
 import java.util.function.Predicate;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
@@ -34,14 +33,13 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.MrnDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
-import uk.gov.hmcts.reform.sscs.ccd.exception.CcdException;
 import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
+import uk.gov.hmcts.reform.sscs.ccd.service.UpdateCcdCaseService;
 import uk.gov.hmcts.reform.sscs.config.CitizenCcdService;
 import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaAppointee;
 import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaCaseWrapper;
 import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaContactDetails;
 import uk.gov.hmcts.reform.sscs.exception.ApplicationErrorException;
-import uk.gov.hmcts.reform.sscs.exception.DuplicateCaseException;
 import uk.gov.hmcts.reform.sscs.idam.IdamService;
 import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
 import uk.gov.hmcts.reform.sscs.idam.UserDetails;
@@ -65,17 +63,18 @@ public abstract class SubmitAppealServiceBase {
     protected final AirLookupService airLookupService;
     protected final RefDataService refDataService;
     protected final VenueService venueService;
+    protected UpdateCcdCaseService updateCcdCaseService;
     protected final boolean caseAccessManagementFeature;
 
-    public SubmitAppealServiceBase(CcdService ccdService,
-                                   CitizenCcdService citizenCcdService,
-                                   IdamService idamService,
-                                   ConvertAIntoBService<SscsCaseData, SessionDraft> convertAIntoBService,
-                                   RegionalProcessingCenterService regionalProcessingCenterService,
-                                   AirLookupService airLookupService,
-                                   RefDataService refDataService,
-                                   VenueService venueService,
-                                   boolean caseAccessManagementFeature) {
+    protected SubmitAppealServiceBase(CcdService ccdService,
+                                      CitizenCcdService citizenCcdService,
+                                      IdamService idamService,
+                                      ConvertAIntoBService<SscsCaseData, SessionDraft> convertAIntoBService,
+                                      RegionalProcessingCenterService regionalProcessingCenterService,
+                                      AirLookupService airLookupService,
+                                      RefDataService refDataService,
+                                      VenueService venueService,
+                                      boolean caseAccessManagementFeature) {
         this.ccdService = ccdService;
         this.citizenCcdService = citizenCcdService;
         this.idamService = idamService;
@@ -86,6 +85,30 @@ public abstract class SubmitAppealServiceBase {
         this.refDataService = refDataService;
         this.venueService = venueService;
     }
+
+    protected SubmitAppealServiceBase(CcdService ccdService,
+                                      CitizenCcdService citizenCcdService,
+                                      IdamService idamService,
+                                      ConvertAIntoBService<SscsCaseData, SessionDraft> convertAIntoBService,
+                                      RegionalProcessingCenterService regionalProcessingCenterService,
+                                      AirLookupService airLookupService,
+                                      RefDataService refDataService,
+                                      VenueService venueService,
+                                      UpdateCcdCaseService updateCcdCaseService,
+                                      boolean caseAccessManagementFeature) {
+        this(ccdService,
+                citizenCcdService,
+                idamService,
+                convertAIntoBService,
+                regionalProcessingCenterService,
+                airLookupService,
+                refDataService,
+                venueService,
+                caseAccessManagementFeature);
+        this.updateCcdCaseService = updateCcdCaseService;
+    }
+
+    public abstract Long submitAppeal(SyaCaseWrapper appeal, String userToken);
 
     public abstract Optional<SaveCaseResult> submitDraftAppeal(String oauth2Token, SyaCaseWrapper appeal, Boolean forceCreate);
 
@@ -110,8 +133,8 @@ public abstract class SubmitAppealServiceBase {
         }
 
         log.info("GET Draft case with CCD Id {} , IDAM Id {} and roles {} ",
-            (caseDetails == null) ? null : caseDetails.getCcdCaseId(), idamTokens.getUserId(),
-            idamTokens.getRoles());
+                (caseDetails == null) ? null : caseDetails.getCcdCaseId(), idamTokens.getUserId(),
+                idamTokens.getRoles());
 
         return (sessionDraft != null) ? Optional.of(sessionDraft) : Optional.empty();
     }
@@ -127,69 +150,15 @@ public abstract class SubmitAppealServiceBase {
         log.info("GET all Draft cases with IDAM Id {} and roles {}", idamTokens.getUserId(), idamTokens.getRoles());
 
         return caseDetailsList.stream()
-            .map(convertAIntoBService::convert)
-            .toList();
+                .map(convertAIntoBService::convert)
+                .toList();
     }
 
-    public Long submitAppeal(SyaCaseWrapper appeal, String userToken) {
-
-        IdamTokens idamTokens = idamService.getIdamTokens();
-
-        Long caseId = appeal.getCcdCaseId() != null ? Long.valueOf(appeal.getCcdCaseId()) : null;
-
-        log.info("Converting sya appeal data to sscs case");
-
-        SscsCaseData caseData = convertAppealToSscsCaseData(appeal);
-
-        EventType event = findEventType(caseData, caseId != null);
-        SscsCaseDetails caseDetails = createOrUpdateCase(caseData, event, idamTokens);
-
-        associateCase(idamTokens, caseDetails, userToken);
-
-        return caseDetails.getId();
-    }
-
-    SscsCaseData convertAppealToSscsCaseData(SyaCaseWrapper appeal) {
-        boolean isIba = appeal.getBenefitType().getCode().equals(Benefit.INFECTED_BLOOD_COMPENSATION.getShortName());
-        String postCode = resolvePostCode(appeal, isIba);
-        RegionalProcessingCenter rpc = regionalProcessingCenterService.getByPostcode(isIba ? postCode : getFirstHalfOfPostcode(postCode));
-
-        SscsCaseData sscsCaseData = rpc == null
-            ? convertSyaToCcdCaseDataV1(appeal, caseAccessManagementFeature)
-            : convertSyaToCcdCaseDataV1(appeal, rpc.getName(), rpc, caseAccessManagementFeature);
-
-        sscsCaseData.setCreatedInGapsFrom(READY_TO_LIST.getId());
-        String processingVenue = airLookupService.lookupAirVenueNameByPostCode(postCode, sscsCaseData.getAppeal().getBenefitType());
-        sscsCaseData.setProcessingVenue(processingVenue);
-
-        if (caseAccessManagementFeature
-            && StringUtils.isNotEmpty(processingVenue)
-            && rpc != null) {
-            String venueEpimsId = venueService.getEpimsIdForVenue(processingVenue);
-            CourtVenue courtVenue = refDataService.getCourtVenueRefDataByEpimsId(venueEpimsId);
-
-            sscsCaseData.setCaseManagementLocation(CaseManagementLocation.builder()
-                .baseLocation(rpc.getEpimsId())
-                .region(courtVenue.getRegionId()).build());
-
-            log.info("Successfully updated case management location details for case {}. Processing venue {}, epimsId {}",
-                appeal.getCcdCaseId(), processingVenue, venueEpimsId);
-        }
-
-        log.info("{} - setting venue name to {}",
-            isIba
-                ? sscsCaseData.getAppeal().getAppellant().getIdentity().getIbcaReference()
-                : maskNino(sscsCaseData.getAppeal().getAppellant().getIdentity().getNino()),
-            sscsCaseData.getProcessingVenue());
-
-        return sscsCaseData;
-    }
-
-    private String maskNino(String nino) {
+    protected String maskNino(String nino) {
         return "XXXX" + (nino == null ? "" : nino.substring(4));
     }
 
-    private EventType findEventType(SscsCaseData caseData, boolean saveAndReturnCase) {
+    protected EventType findEventType(SscsCaseData caseData, boolean saveAndReturnCase) {
         Appeal appeal = caseData.getAppeal();
         MrnDetails mrnDetails = appeal.getMrnDetails();
         if (mrnDetails != null && mrnDetails.getMrnDate() != null) {
@@ -200,25 +169,25 @@ public abstract class SubmitAppealServiceBase {
         } else {
             Identity identity = appeal.getAppellant().getIdentity();
             log.info("Moving case for {} {} to incomplete due to MRN Details {} present and MRN Date {} present",
-                caseData.isIbcCase() ? "Ibca Reference" : "NINO",
-                caseData.isIbcCase() ? identity.getIbcaReference() : maskNino(identity.getNino()),
-                (mrnDetails != null ? "" : "not"),
-                (mrnDetails != null && mrnDetails.getMrnDate() != null ? "" : "not"));
+                    caseData.isIbcCase() ? "Ibca Reference" : "NINO",
+                    caseData.isIbcCase() ? identity.getIbcaReference() : maskNino(identity.getNino()),
+                    (mrnDetails != null ? "" : "not"),
+                    (mrnDetails != null && mrnDetails.getMrnDate() != null ? "" : "not"));
             return saveAndReturnCase ? DRAFT_TO_INCOMPLETE_APPLICATION : INCOMPLETE_APPLICATION_RECEIVED;
         }
     }
 
-    private Pair<List<SscsCaseDetails>, SscsCaseDetails> getMatchedNinoCasesAndCaseDetails(SscsCaseData caseData,
-                                                                                           IdamTokens idamTokens,
-                                                                                           String nino,
-                                                                                           String ibcaReference) {
+    protected Pair<List<SscsCaseDetails>, SscsCaseDetails> getMatchedNinoCasesAndCaseDetails(SscsCaseData caseData,
+                                                                                             IdamTokens idamTokens,
+                                                                                             String nino,
+                                                                                             String ibcaReference) {
         SscsCaseDetails caseDetails = null;
         List<SscsCaseDetails> matchedByNinoCases = List.of();
         if (!caseData.isIbcCase()) {
             matchedByNinoCases = getMatchedCases(nino, idamTokens);
             if (!matchedByNinoCases.isEmpty()) {
                 log.info("Found " + matchedByNinoCases.size() + " matching cases for Nino "
-                    + maskNino(nino) + " before filtering non exact matches");
+                        + maskNino(nino) + " before filtering non exact matches");
             } else {
                 log.info("No matching cases for Nino {}", maskNino(nino));
             }
@@ -229,95 +198,24 @@ public abstract class SubmitAppealServiceBase {
         return Pair.of(matchedByNinoCases, caseDetails);
     }
 
-    private SscsCaseDetails createOrUpdateCase(SscsCaseData caseData, EventType eventType, IdamTokens idamTokens) {
-        SscsCaseDetails caseDetails = null;
-        String benefitShortName = caseData.getAppeal().getBenefitType().getCode();
-        String nino = caseData.getAppeal().getAppellant().getIdentity().getNino();
-        String ibcaReference = caseData.getAppeal().getAppellant().getIdentity().getIbcaReference();
-        try {
-            Pair<List<SscsCaseDetails>, SscsCaseDetails> matchedByNinoCasesCaseDetails =
-                getMatchedNinoCasesAndCaseDetails(caseData, idamTokens, nino, ibcaReference);
-            List<SscsCaseDetails> matchedByNinoCases = matchedByNinoCasesCaseDetails.getLeft();
-            caseDetails = matchedByNinoCasesCaseDetails.getRight();
-            if (caseDetails == null) {
-                if (!matchedByNinoCases.isEmpty()) {
-                    log.info("Found " + matchedByNinoCases.size() + " matching cases for Nino "
-                        + maskNino(nino));
-                    caseData = addAssociatedCases(caseData, matchedByNinoCases);
-                }
-
-                log.info("About to attempt creating case or updating draft case in CCD with event {} for benefit type {} and event {} and isScottish {} and languagePreference {}",
-                    eventType,
-                    benefitShortName,
-                    eventType,
-                    caseData.getIsScottishCase(),
-                    caseData.getLanguagePreference().getCode());
-
-                if (eventType == DRAFT_TO_VALID_APPEAL_CREATED || eventType == DRAFT_TO_INCOMPLETE_APPLICATION || eventType == DRAFT_TO_NON_COMPLIANT) {
-                    caseData.setCaseCreated(LocalDate.now().toString());
-
-                    caseDetails = ccdService.updateCase(caseData,
-                        Long.valueOf(caseData.getCcdCaseId()),
-                        eventType.getCcdType(),
-                        "SSCS - new case created",
-                        "Created SSCS case from Submit Your Appeal online draft with event " + eventType.getCcdType(),
-                        idamTokens);
-
-                    log.info("Case {} successfully converted from Draft to SSCS case in CCD for benefit type {} with event {}",
-                        caseDetails.getId(),
-                        benefitShortName,
-                        eventType);
-                } else {
-                    caseDetails = ccdService.createCase(caseData,
-                        eventType.getCcdType(),
-                        "SSCS - new case created",
-                        "Created SSCS case from Submit Your Appeal online with event " + eventType.getCcdType(),
-                        idamTokens);
-                    log.info("Case {} successfully created in CCD for benefit type {} with event {}",
-                        caseDetails.getId(),
-                        benefitShortName,
-                        eventType);
-                }
-                return caseDetails;
-            }
-        } catch (Exception e) {
-            String caseId = caseDetails != null ? caseDetails.getId().toString() : "";
-            String referenceName = caseData.isIbcCase() ? "IBCA Reference" : "Nino";
-            String referenceValue = caseData.isIbcCase() ? ibcaReference : maskNino(nino);
-            throw new CcdException(
-                String.format("Error found in the creating case process for case with Id - %s"
-                        + " and %s - %s and Benefit type - %s and exception: %s",
-                    caseId, referenceName, referenceValue,
-                    benefitShortName, e.getMessage()), e);
-        }
-
-        log.info("Duplicate case {} found for Nino {} and benefit type {}. "
-                + "No need to continue with post create case processing.",
-            caseDetails.getId(), maskNino(nino),
-            benefitShortName);
-        throw new DuplicateCaseException(
-            String.format("An appeal has already been submitted, for that decision date %s ",
-                caseData.getAppeal().getMrnDetails().getMrnDate()));
-    }
-
-    private void associateCase(IdamTokens idamTokens,
-                               SscsCaseDetails caseDetails,
-                               String userToken) {
+    protected void associateCase(IdamTokens idamTokens,
+                                 SscsCaseDetails caseDetails,
+                                 String userToken) {
         if (null != caseDetails && StringUtils.isNotEmpty(userToken)) {
             citizenCcdService.associateCaseToCitizen(getUserTokens(userToken), caseDetails.getId(), idamTokens);
         }
     }
 
-    private String resolvePostCode(SyaCaseWrapper appeal, boolean isIba) {
+    protected String resolvePostCode(SyaCaseWrapper appeal, boolean isIba) {
         Boolean inMainlandUk = appeal.getContactDetails().getInMainlandUk();
         if (isIba && inMainlandUk != null && inMainlandUk.equals(Boolean.FALSE)) {
             return appeal.getAppellant().getContactDetails().getPortOfEntry();
         } else if (Boolean.TRUE.equals(appeal.getIsAppointee())) {
             return Optional.ofNullable(appeal.getAppointee())
-                .map(SyaAppointee::getContactDetails)
-                .map(SyaContactDetails::getPostCode)
-                .filter(appointeePostCode -> !StringUtils.isEmpty(appointeePostCode))
-                .orElse(null);
+                    .map(SyaAppointee::getContactDetails)
+                    .map(SyaContactDetails::getPostCode)
+                    .filter(appointeePostCode -> !StringUtils.isEmpty(appointeePostCode))
+                    .orElse(null);
         } else {
             return appeal.getAppellant().getContactDetails().getPostCode();
         }
@@ -339,42 +237,39 @@ public abstract class SubmitAppealServiceBase {
         return ccdService.findCaseBy("data.appeal.appellant.identity.nino", nino, idamTokens);
     }
 
-    private Predicate<SscsCaseDetails> createNinoAndBenefitTypeAndMrnDatePredicate(SscsCaseData caseData) {
+    protected Predicate<SscsCaseDetails> createNinoAndBenefitTypeAndMrnDatePredicate(SscsCaseData caseData) {
         return c -> c.getData().getAppeal().getAppellant().getIdentity() != null
-            && c.getData().getAppeal().getAppellant().getIdentity().getNino().equalsIgnoreCase(caseData.getAppeal().getAppellant().getIdentity().getNino())
-            && c.getData().getAppeal().getBenefitType() != null
-            && c.getData().getAppeal().getBenefitType().getCode().equals(caseData.getAppeal().getBenefitType().getCode())
-            && c.getData().getAppeal().getMrnDetails().getMrnDate() != null
-            && c.getData().getAppeal().getMrnDetails().getMrnDate().equalsIgnoreCase(caseData.getAppeal().getMrnDetails().getMrnDate());
+                && c.getData().getAppeal().getAppellant().getIdentity().getNino().equalsIgnoreCase(caseData.getAppeal().getAppellant().getIdentity().getNino())
+                && c.getData().getAppeal().getBenefitType() != null
+                && c.getData().getAppeal().getBenefitType().getCode().equals(caseData.getAppeal().getBenefitType().getCode())
+                && c.getData().getAppeal().getMrnDetails().getMrnDate() != null
+                && c.getData().getAppeal().getMrnDetails().getMrnDate().equalsIgnoreCase(caseData.getAppeal().getMrnDetails().getMrnDate());
     }
 
-    protected SscsCaseData addAssociatedCases(SscsCaseData caseData, List<SscsCaseDetails> matchedByNinoCases) {
-        log.info("Adding " + matchedByNinoCases.size() + " associated cases for case id {}", caseData.getCcdCaseId());
+    public void addAssociatedCases(final SscsCaseData caseData, List<SscsCaseDetails> matchedByNinoCases) {
+        log.info("Adding {} associated cases for case id {}", matchedByNinoCases.size(), caseData.getCcdCaseId());
 
-        List<CaseLink> associatedCases = new ArrayList<>();
+        List<CaseLink> caseLinks = matchedByNinoCases.stream().map(sscsCaseDetails -> CaseLink.builder().value(
+                        CaseLinkDetails.builder().caseReference(sscsCaseDetails.getId().toString()).build()).build())
+                .toList();
 
-        for (SscsCaseDetails sscsCaseDetails : matchedByNinoCases) {
-            log.info("Linking case " + sscsCaseDetails.getId().toString());
-            associatedCases.add(CaseLink.builder().value(
-                CaseLinkDetails.builder().caseReference(sscsCaseDetails.getId().toString()).build()).build());
-        }
-
-        if (!matchedByNinoCases.isEmpty()) {
-            return caseData.toBuilder().associatedCase(associatedCases).linkedCasesBoolean("Yes").build();
+        if (caseLinks.isEmpty()) {
+            caseData.setLinkedCasesBoolean("No");
         } else {
-            return caseData.toBuilder().linkedCasesBoolean("No").build();
+            caseData.setAssociatedCase(caseLinks);
+            caseData.setLinkedCasesBoolean("Yes");
         }
     }
 
     protected IdamTokens getUserTokens(String oauth2Token) {
         UserDetails userDetails = idamService.getUserDetails(oauth2Token);
         return IdamTokens.builder()
-            .idamOauth2Token(oauth2Token)
-            .serviceAuthorization(idamService.generateServiceAuthorization())
-            .userId(userDetails.getId())
-            .roles(userDetails.getRoles())
-            .email(userDetails.getEmail())
-            .build();
+                .idamOauth2Token(oauth2Token)
+                .serviceAuthorization(idamService.generateServiceAuthorization())
+                .userId(userDetails.getId())
+                .roles(userDetails.getRoles())
+                .email(userDetails.getEmail())
+                .build();
     }
 
     protected boolean hasValidCitizenRole(IdamTokens idamTokens) {
@@ -388,12 +283,49 @@ public abstract class SubmitAppealServiceBase {
     protected void logError(SyaCaseWrapper appeal, IdamTokens idamTokens) {
         if (nonNull(appeal.getAppellant().getNino())) {
             log.error("The case data has been altered outside of this transaction for case with nino {} and idam id {}",
-                maskNino(appeal.getAppellant().getNino()),
-                idamTokens.getUserId());
+                    maskNino(appeal.getAppellant().getNino()),
+                    idamTokens.getUserId());
         } else {
             log.error("The case data has been altered outside of this transaction for idam id {}",
-                idamTokens.getUserId());
+                    idamTokens.getUserId());
         }
+    }
+
+
+    protected SscsCaseData convertAppealToSscsCaseData(SyaCaseWrapper appeal,
+                                                       Function<RegionalProcessingCenter, SscsCaseData> getCaseData) {
+        boolean isIba = appeal.getBenefitType().getCode().equals(Benefit.INFECTED_BLOOD_COMPENSATION.getShortName());
+        String postCode = resolvePostCode(appeal, isIba);
+        RegionalProcessingCenter rpc = regionalProcessingCenterService.getByPostcode(isIba ? postCode : getFirstHalfOfPostcode(postCode));
+
+
+        SscsCaseData sscsCaseData = getCaseData.apply(rpc);
+
+        sscsCaseData.setCreatedInGapsFrom(READY_TO_LIST.getId());
+        String processingVenue = airLookupService.lookupAirVenueNameByPostCode(postCode, sscsCaseData.getAppeal().getBenefitType());
+        sscsCaseData.setProcessingVenue(processingVenue);
+
+        if (caseAccessManagementFeature
+                && StringUtils.isNotEmpty(processingVenue)
+                && rpc != null) {
+            String venueEpimsId = venueService.getEpimsIdForVenue(processingVenue);
+            CourtVenue courtVenue = refDataService.getCourtVenueRefDataByEpimsId(venueEpimsId);
+
+            sscsCaseData.setCaseManagementLocation(CaseManagementLocation.builder()
+                    .baseLocation(rpc.getEpimsId())
+                    .region(courtVenue.getRegionId()).build());
+
+            log.info("Successfully updated case management location details for case {}. Processing venue {}, epimsId {}",
+                    appeal.getCcdCaseId(), processingVenue, venueEpimsId);
+        }
+
+        log.info("{} - setting venue name to {}",
+                isIba
+                        ? sscsCaseData.getAppeal().getAppellant().getIdentity().getIbcaReference()
+                        : maskNino(sscsCaseData.getAppeal().getAppellant().getIdentity().getNino()),
+                sscsCaseData.getProcessingVenue());
+
+        return sscsCaseData;
     }
 
 }
