+++ src/test/resources/sya/allDetailsNonSaveAndReturnCcdSscs5.json
@@ -278,5 +278,6 @@
   "dwpIsOfficerAttending":  "No",
   "poAttendanceConfirmed": "No",
   "tribunalDirectPoToAttend": "No",
-  "hearingRoute":"gaps"
+  "hearingRoute":"gaps",
+  "linkedCasesBoolean":"No"
 }
