+++ src/test/resources/sya/allDetailsNonSaveAndReturnWithInterlocCcd.json
@@ -281,5 +281,6 @@
   "dwpIsOfficerAttending":  "No",
   "poAttendanceConfirmed": "No",
   "tribunalDirectPoToAttend": "No",
-  "hearingRoute":"gaps"
+  "hearingRoute":"gaps",
+  "linkedCasesBoolean": "No"
 }
